Remove changes to readme, move files to src
