+++ src/browser.d.ts
@@ -1,2 +1,2 @@
 // NOTE: dotEnvPath option will be ignored when using browsers.
-export * from './lib/envalid.d.ts'
+export * from './envalid.d.ts'
