+++ src/extend.js
