+++ src/strictProxy.js
