+++ tests/test_basics.js
@@ -1,5 +1,5 @@
 const { createGroup, assert } = require('painless')
-const { cleanEnv, EnvError, EnvMissingError, str, num, testOnly } = require('..')
+const { cleanEnv, EnvError, EnvMissingError, str, num, testOnly } = require('../src')
 const { assertPassthrough } = require('./utils')
 const test = createGroup()
 const makeSilent = { reporter: null }
